integrate graph setup with mockup
